reogranized folder structure to reflect size of project
