+++ test_adegorical.py
